Moved benchmark test into its own file
